preprocessing/requirements.txt
fastapi==0.115.12
uvicorn==0.34.2
pandas==2.2.3
SQLAlchemy==2.0.40
psycopg2-binary==2.9.9
python-dotenv==1.1.0